NCI-Agency/anet#822: Move GraphQL helper classes to separate package
