+++ src/test/java/mil/dds/anet/test/resources/utils/GraphQLClient.java
@@ -1,4 +1,4 @@
-package mil.dds.anet.test.resources;
+package mil.dds.anet.test.resources.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
